Add anchor text generation code by Michael Moore
This code generates a dataset containing for each article all titles
which are used in internal wiki links to point to this article. This
data is later-on used to create a surface form dataset which could be
interesting, e.g., for NLP applications. The actual code for the surface
form dataset is not yet included in the framework.