dummy edit to add comment about last change
Concerning the last change to UriPolicy.scala: I was blind - genericDomain was used. But anyway, the extraction.properties files don't use the abilty to specify the generic domain, I didn't find any uses in the scripts, and it shouldn't be necessary for normal use of the framework - changing a URI from specific to generic is the last thing that is done before it is serialized.